refactor(pipeline): extract hub SSH helper to ssh_hub.py (#568)
ssh.py was 83 lines, over the 80-line pipeline cap. Extracted the
hub-transport helper into a new sibling stages/common/ssh_hub.py:

- ssh_hub.py (28) — hub_ssh_run() (ProxyJump to hub VM) + the
  saconsole_ssh_run backward-compat alias, moved verbatim
- ssh.py (83→69) — keeps the target-VM transport helpers (ssh_run,
  scp_to_vm, rsync_to_vm); re-exports hub_ssh_run + saconsole_ssh_run
  from ssh_hub so existing import sites stay unchanged

Facade chosen over repointing the lone hub importer (tls_cert.py): a
clean repoint would split its combined import and grow it 63→64,
which the size ratchet forbids. tls_cert.py stays byte-identical.
Dropped now-unused DEFAULT_HYPERVISOR / HUB_VIRBR0_IP imports from ssh.py.

Pure refactor, no behavioral change: import smoke test confirms all
symbols resolve to the same objects; live e2e stage_5 (3/3) and
stage_6 (6/6) against dev02 exercise ssh_run + scp_to_vm.

fixes #568
