+++ tools/size_baselines.json
@@ -125,7 +125,8 @@
   "tools/pipeline/stages/common/connectivity.py": 18,
   "tools/pipeline/stages/common/known_hosts.py": 69,
   "tools/pipeline/stages/common/log_format.py": 22,
-  "tools/pipeline/stages/common/ssh.py": 83,
+  "tools/pipeline/stages/common/ssh.py": 69,
+  "tools/pipeline/stages/common/ssh_hub.py": 28,
   "tools/pipeline/stages/common/types.py": 57,
   "tools/pipeline/stages/common/verify_ansible_output.py": 46,
   "tools/pipeline/stages/common/verify_args.py": 64,
